+++ src/lib/cdm/typescript/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @dotdm/cdm
 
+## 0.6.26
+
+### Patch Changes
+
+- Updated dependencies [3b23b96]
+  - @parity/cdm-env@2.4.0
+  - @parity/cdm-builder@5.0.1
+
 ## 0.6.25
 
 ### Patch Changes